requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asttokens==3.0.0
attrs==24.2.0
av==13.1.0
bitsandbytes==0.44.1
black==24.10.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cut-cross-entropy==24.11.4
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
dill==0.3.8
distlib==0.3.9
docstring_parser==0.16
einops==0.8.0
executing==2.1.0
filelock==3.16.1
flash-attn==2.7.0.post2
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
hf_transfer==0.1.8
huggingface-hub==0.26.2
identify==2.6.2
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.1.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
peft==0.13.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.4.0
shtab==1.7.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
trl==0.12.1
typeguard==4.4.1
typing_extensions==4.12.2
tyro==0.9.2
tzdata==2024.2
unsloth @ git+https://github.com/unslothai/unsloth.git@8558bc92b06f9128499484ef737fa71b966ffc23
unsloth_zoo==2024.11.8
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
wheel==0.45.1
xformers==0.0.28.post3
xxhash==3.5.0
yarl==1.17.1
